Finalizing logic of quiz

+++ app/src/main/java/com/mrkevin574/quizsabana/presentation/Screens.kt
@@ -5,7 +5,7 @@ import com.mrkevin574.quizsabana.util.SCORE_KEY
 sealed class Screens(val route : String)
 {
     object FinalScreen : Screens("final_screen/{$SCORE_KEY}"){
-        fun passScore(score : Int) = "final_screen/$score"
+        fun passScore(score : Float) = "final_screen/$score"
     }
     object QuizScreen : Screens("quiz_screen")
     object ScoreScreen : Screens("score_screen")
